// src/features/meet/types.ts
export interface JitsiMeetAPIOptions {
  roomName?: string;

  width?: number;

  height?: number;

  parentNode?: ReactNode | null;

  configOverwrite?: ConfigOptions;

  interfaceConfigOverwrite?: InterfaceConfigOptions;

  noSSL?: boolean;

  jwt?: string;

  onLoad?: () => void;

  invites?: object;

  devices?: {
    audioInput?: string;
    audioOutput?: string;
    videoInput?: string;
  };

  userInfo?: {
    email?: string;
  };
}